+++ cli/cli_helpers/persona_ingest.py
@@ -1,4 +1,0 @@
-from .persona_seeder import cmd_persona_load as _cmd_persona_load
-
-def cmd_persona_ingest(args):
-    return _cmd_persona_load(args)